feat: move all core highlights into single file
I'm inspried by other plugins such as rose-pine and tokyonight to
keep all of the highlights for plugins (other than external
integrations) in a single file. That single file can be cached
and sped up better, and it's a little easier to maintain. Most people
are not going to be enabling plugins like Catppuccin's setup model.